requirements.txt
streamlit
torch
torchvision
requests
pillow
numpy